# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Router public/admin API (JWT/UCAN) under /api/v1.",
        "title": "Router API",
        "contact": {},
        "version": "v1"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/admin/log": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List logs (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page index",
                        "name": "p",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Log type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Start timestamp (unix)",
                        "name": "start_timestamp",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "End timestamp (unix)",
                        "name": "end_timestamp",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "username",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Token name",
                        "name": "token_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Model name",
                        "name": "model_name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Channel ID",
                        "name": "channel",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/docs.StandardResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Delete history logs (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Target timestamp (unix)",
                        "name": "target_timestamp",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/docs.StandardResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/log/search": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Search logs (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Keyword",
                        "name": "keyword",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/docs.StandardResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/log/stat": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Log stats (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Log type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Start timestamp (unix)",
                        "name": "start_timestamp",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "End timestamp (unix)",
                        "name": "end_timestamp",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Token name",
                        "name": "token_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "username",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Model name",
                        "name": "model_name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Channel ID",
                        "name": "channel",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/docs.StandardResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/public/log/self": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "summary": "List user logs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page index",
                        "name": "p",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Log type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Start timestamp (unix)",
                        "name": "start_timestamp",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "End timestamp (unix)",
                        "name": "end_timestamp",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Token name",
                        "name": "token_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Model name",
                        "name": "model_name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/docs.StandardResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/public/log/self/search": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "summary": "Search user logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Keyword",
                        "name": "keyword",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/docs.StandardResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/public/log/self/stat": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "summary": "Log stats for current user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Log type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Start timestamp (unix)",
                        "name": "start_timestamp",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "End timestamp (unix)",
                        "name": "end_timestamp",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Token name",
                        "name": "token_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Model name",
                        "name": "model_name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Channel ID",
                        "name": "channel",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/docs.StandardResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "docs.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
...
        "docs.StandardResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}